requirements.txt
selenium
webdriver-manager
behave
openpyxl
allure-behave
allure-python-commons